feat: use calimero client js blob api
