+++ apps/blob-test/frontend/src/api/chatApi.ts
@@ -1,21 +1,61 @@
-import { 
-  ApiResponse, 
-  RpcQueryParams, 
-  rpcClient
-} from '@calimero-network/calimero-client';
-import { 
-  ChatApi, 
-  BlobUploadResponse, 
-  BlobMetadataResponse,
-  SendMessageRequest,
-  Message,
-  Attachment,
-  ChatStats,
-  ChatMethod
-} from '../blobApi';
-import { getAuthConfig } from '@calimero-network/calimero-client';
+import { ApiResponse, RpcQueryParams, rpcClient, getAuthConfig } from '@calimero-network/calimero-client';
 
-const BASE_URL = process.env.REACT_APP_NODE_URL || 'http://localhost:2428';
+// Chat API interfaces for JSON RPC functionality
+export interface Message {
+  id: number;
+  sender: string;
+  text: string;
+  attachments: Attachment[];
+  timestamp: number;
+}
+
+export interface Attachment {
+  original_name: string;
+  original_blob_id: string;
+  original_size: number;
+  compressed_blob_id: string;
+  compressed_size: number;
+  content_type?: string;
+  compression_ratio: number;
+}
+
+export interface SendMessageRequest {
+  sender: string;
+  text: string;
+  attachment_blob_ids: string[];
+  attachment_names: string[];
+  attachment_sizes: number[];
+  attachment_content_types: (string | null)[];
+}
+
+export interface ChatStats extends Record<string, number> {
+  total_messages: number;
+  total_attachments: number;
+  total_original_size_bytes: number;
+  total_compressed_size_bytes: number;
+  compression_savings_percent: number;
+}
+
+// Method names that match the Rust backend  
+export enum ChatMethod {
+  SEND_MESSAGE = 'send_message',
+  GET_MESSAGES = 'get_messages',
+  GET_MESSAGE = 'get_message',
+  GET_DECOMPRESSED_BLOB_ID = 'get_decompressed_blob_id',
+  GET_STATS = 'get_stats',
+  CLEAR_MESSAGES = 'clear_messages',
+}
+
+// File upload interface for the UI
+export interface FileUpload {
+  file: File;
+  blob_id?: string;
+  uploading: boolean;
+  uploaded: boolean;
+  progress: number;
+  error?: string;
+}
+
 const RequestConfig = { timeout: 30000 };
 
 function getErrorMessage(error: any): string {
@@ -25,181 +65,9 @@ function getErrorMessage(error: any): string {
   return 'An unexpected error occurred';
 }
 
-export class ChatApiDataSource implements ChatApi {
+// Chat API implementation using JSON RPC
+export class ChatApi {
   
-  // REST API Methods for file upload/download
-
-  async uploadBlob(
-    file: File, 
-    onProgress?: (progress: number) => void, 
-    expectedHash?: string
-  ): Promise<ApiResponse<BlobUploadResponse>> {
-    return this.uploadBlobRaw(file, onProgress, expectedHash);
-  }
-
-  async uploadBlobRaw(
-    file: File, 
-    onProgress?: (progress: number) => void, 
-    expectedHash?: string
-  ): Promise<ApiResponse<BlobUploadResponse>> {
-    // Read file as ArrayBuffer for raw binary upload
-    const fileArrayBuffer = await file.arrayBuffer();
-
-    return new Promise((resolve) => {
-      const xhr = new XMLHttpRequest();
-
-      xhr.upload.addEventListener('progress', (event) => {
-        if (event.lengthComputable && onProgress) {
-          const progress = (event.loaded / event.total) * 100;
-          onProgress(progress);
-        }
-      });
-
-      xhr.addEventListener('load', () => {
-        try {
-          if (xhr.status === 200) {
-            const response = JSON.parse(xhr.responseText);
-            resolve({
-              data: response,
-              error: null,
-            });
-          } else {
-            const errorResponse = JSON.parse(xhr.responseText);
-            resolve({
-              data: null,
-              error: {
-                code: xhr.status,
-                message: errorResponse.error || `HTTP ${xhr.status}: ${xhr.statusText}`,
-              },
-            });
-          }
-        } catch (error) {
-          resolve({
-            data: null,
-            error: {
-              code: xhr.status || 500,
-              message: error instanceof Error ? error.message : 'Failed to parse response',
-            },
-          });
-        }
-      });
-
-      xhr.addEventListener('error', () => {
-        resolve({
-          data: null,
-          error: {
-            code: 500,
-            message: 'Network error occurred during upload',
-          },
-        });
-      });
-
-      // Build URL with query parameters
-      let url = `${BASE_URL}/admin-api/blobs/upload-raw`;
-      if (expectedHash) {
-        url += `?hash=${encodeURIComponent(expectedHash)}`;
-      }
-
-      xhr.open('POST', url);
-      xhr.setRequestHeader('Content-Type', 'application/octet-stream');
-      xhr.send(fileArrayBuffer);
-    });
-  }
-
-  async downloadBlob(blobId: string): Promise<Blob> {
-    const response = await fetch(`${BASE_URL}/admin-api/blobs/${blobId}`);
-    
-    if (!response.ok) {
-      throw new Error(`Failed to download blob: ${response.status} ${response.statusText}`);
-    }
-    
-    return response.blob();
-  }
-
-  async getBlobMetadata(blobId: string): Promise<ApiResponse<BlobMetadataResponse>> {
-    try {
-      const response = await fetch(`${BASE_URL}/admin-api/blobs/${blobId}/info`);
-      
-      if (!response.ok) {
-        return {
-          data: null,
-          error: {
-            code: response.status,
-            message: `HTTP ${response.status}: ${response.statusText}`,
-          },
-        };
-      }
-      
-      const data = await response.json();
-      return {
-        data,
-        error: null,
-      };
-    } catch (error) {
-      console.error('getBlobMetadata failed:', error);
-      return {
-        error: {
-          code: 500,
-          message: error instanceof Error ? error.message : 'An unexpected error occurred',
-        },
-      };
-    }
-  }
-
-  // Helper methods for attachment handling
-
-  async getDecompressedBlobId(compressedBlobId: string): ApiResponse<string> {
-    try {
-      const config = getAuthConfig();
-
-      if (!config || !config.contextId || !config.executorPublicKey) {
-        return {
-          data: null,
-          error: {
-            code: 500,
-            message: 'Authentication configuration not found',
-          },
-        };
-      }
-
-      const params: RpcQueryParams<{ compressed_blob_id_str: string }> = {
-        contextId: config.contextId,
-        method: ChatMethod.GET_DECOMPRESSED_BLOB_ID,
-        argsJson: { compressed_blob_id_str: compressedBlobId },
-        executorPublicKey: config.executorPublicKey,
-      };
-
-      const response = await rpcClient.execute<{ compressed_blob_id_str: string }, string>(
-        params, 
-        RequestConfig
-      );
-
-      if (response?.error) {
-        return {
-          error: {
-            code: response.error.code ?? 500,
-            message: getErrorMessage(response.error)
-          }
-        };
-      }
-
-      return {
-        data: response.result.output as string,
-        error: null,
-      };
-    } catch (error) {
-      console.error('getDecompressedBlobId failed:', error);
-      return {
-        error: {
-          code: 500,
-          message: error instanceof Error ? error.message : 'An unexpected error occurred',
-        },
-      };
-    }
-  }
-
-  // JSON RPC Methods for Chat Functionality
-
   async sendMessage(request: SendMessageRequest): ApiResponse<number> {
     try {
       const config = getAuthConfig();
@@ -341,6 +209,56 @@ export class ChatApiDataSource implements ChatApi {
       };
     } catch (error) {
       console.error('getMessage failed:', error);
+      return {
+        error: {
+          code: 500,
+          message: error instanceof Error ? error.message : 'An unexpected error occurred',
+        },
+      };
+    }
+  }
+
+  async getDecompressedBlobId(compressedBlobId: string): ApiResponse<string> {
+    try {
+      const config = getAuthConfig();
+
+      if (!config || !config.contextId || !config.executorPublicKey) {
+        return {
+          data: null,
+          error: {
+            code: 500,
+            message: 'Authentication configuration not found',
+          },
+        };
+      }
+
+      const params: RpcQueryParams<{ compressed_blob_id_str: string }> = {
+        contextId: config.contextId,
+        method: ChatMethod.GET_DECOMPRESSED_BLOB_ID,
+        argsJson: { compressed_blob_id_str: compressedBlobId },
+        executorPublicKey: config.executorPublicKey,
+      };
+
+      const response = await rpcClient.execute<{ compressed_blob_id_str: string }, string>(
+        params, 
+        RequestConfig
+      );
+
+      if (response?.error) {
+        return {
+          error: {
+            code: response.error.code ?? 500,
+            message: getErrorMessage(response.error)
+          }
+        };
+      }
+
+      return {
+        data: response.result.output as string,
+        error: null,
+      };
+    } catch (error) {
+      console.error('getDecompressedBlobId failed:', error);
       return {
         error: {
           code: 500,
